backend/foodgram/requirements.txt
asgiref==3.6.0
certifi==2022.12.7
cffi==1.15.1
charset-normalizer==3.1.0
click==8.1.3
colorama==0.4.6
cryptography==40.0.2
defusedxml==0.7.1
Django==3.2
django-filter==23.1
django-templated-mail==1.1.1
django-utils-six==2.0
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
djoser==2.2.0
Flask==2.2.3
idna==3.4
importlib-metadata==6.6.0
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.2
oauthlib==3.2.2
Pillow==9.5.0
pycparser==2.21
PyJWT==2.6.0
python3-openid==3.2.0
pytz==2023.3
requests==2.28.2
requests-oauthlib==1.3.1
social-auth-app-django==5.2.0
social-auth-core==4.4.2
sqlparse==0.4.4
tzdata==2023.3
urllib3==1.26.15
webcolors==1.13
Werkzeug==2.2.3
zipp==3.15.0
